requirements.txt
keras==2.8.0
tensorflow==2.8.0
pyarrow==4.0.0
jsonschema==0.7
nltk==3.6.5
requests==2.26.0
gunicorn==20.1.0
pandas==1.3.5
pyyaml==6.0
flask==2.1.0